+++ jsonargparse_tests/test_cli.py
@@ -6,7 +6,7 @@ import yaml
 from contextlib import redirect_stderr, redirect_stdout
 from io import StringIO
 from jsonargparse import ArgumentParser, capture_parser, CLI, lazy_instance
-from jsonargparse.optionals import docstring_parser_support, import_docstring_parse, ruyaml_support
+from jsonargparse.optionals import docstring_parser_support, ruyaml_support
 from jsonargparse.typing import final
 from jsonargparse_tests.base import mock_module, TempDirTestCase
 
@@ -84,8 +84,8 @@ class CLITests(unittest.TestCase):
             self.assertIn('Description of method1:', out.getvalue())
 
             with unittest.mock.patch('docstring_parser.parse') as docstring_parse:
-                DocstringParseError = import_docstring_parse('test_docstring_parse_fail', True)[1]
-                docstring_parse.side_effect = DocstringParseError
+                from docstring_parser import ParseError
+                docstring_parse.side_effect = ParseError
                 out = StringIO()
                 with redirect_stdout(out), self.assertRaises(SystemExit):
                     CLI(Class1, args=['x', 'method1', '--help'])